requirements.txt
gspread
pandas
matplotlib
seaborn
google-auth
google-auth-oauthlib